requirements/backend.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langchain>=0.0.295
langchain-ollama>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
spacy>=3.7.0
Wikipedia-API>=0.6.0
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.23.0
pydantic>=2.0.0
cloudflared>=0.1.0
